Changes needed to compile with Visual Studio 2013. Fixed minor warnings in freeregions.cc and probpipeline.cc. More careful handling of static vs shared library generation depending on the STATICON setting. VS2013 notes added to NERDS_ONLY.txt. Note that multovl was built w/o Bamtools support, only the BED/GFF input works. Release build only, no unit tests run.

+++ src/file/prob/freeregions.cc
@@ -98,7 +98,7 @@ const Region& FreeRegions::select_free_region(UniformGen& rng, unsigned int minl
 {
     unsigned int idx;
     do {
-        float rnd = rng(); // 0.0 .. 1.0
+        float rnd = static_cast<float>(rng()); // 0.0 .. 1.0
         std::vector<float>::const_iterator roulit = 
             std::lower_bound(_roulette_sectors.begin(), _roulette_sectors.end(), rnd);
         idx = (roulit - _roulette_sectors.begin());
